refactor views to hooks for cleaner architecture
