+++ .env
@@ -1,0 +1,1 @@
+VITE_DATA_SOURCE=api
